backend/report_backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-caching==2.3.0
yfinance==0.2.54
pandas==2.2.3
numpy==2.0.2
python-dotenv==1.0.1
requests==2.32.3